refactor(cmd): unify should_read_stdin / should_use_stdin Two identical functions (!stdin().is_terminal() && args.is_empty()) existed under different names. Consolidate into cmd::should_read_stdin.
