+++ crates/rskim/src/cmd/infra/mod.rs
@@ -11,7 +11,6 @@ pub(crate) mod curl;
 pub(crate) mod gh;
 pub(crate) mod wget;
 
-use std::io::IsTerminal;
 use std::process::ExitCode;
 
 use super::{extract_show_stats, run_parsed_command_with_mode, ParsedCommandConfig};
@@ -143,16 +142,6 @@ pub(crate) struct InfraToolConfig<'a> {
     pub install_hint: &'a str,
 }
 
-/// Detect whether to read from piped stdin.
-///
-/// Returns `true` when stdin is not a terminal AND `args` is empty — the same
-/// heuristic used across all infra tool dispatchers.  Centralised here so that
-/// [`run_infra_tool`] and the streaming sub-commands (`gh run watch`, `gh api`)
-/// share a single implementation.
-pub(crate) fn should_use_stdin(args: &[String]) -> bool {
-    !std::io::stdin().is_terminal() && args.is_empty()
-}
-
 /// Execute an infra tool, parse its output, and emit the result.
 ///
 /// This is the single implementation shared by all infra parsers, handling both
@@ -168,7 +157,7 @@ pub(crate) fn run_infra_tool(
     let mut cmd_args = args.to_vec();
     prepare_args(&mut cmd_args);
 
-    let use_stdin = should_use_stdin(args);
+    let use_stdin = super::should_read_stdin(args);
 
     run_parsed_command_with_mode(
         ParsedCommandConfig {
